requirements.txt
torch==1.6.0
lark-parser~=0.9.0
networkx~=2.5